+++ src/test/java/org/zalando/nakadi/service/StorageServiceTest.java
@@ -3,10 +3,12 @@ package org.zalando.nakadi.service;
 import org.json.JSONObject;
 import org.junit.Before;
 import org.junit.Test;
+import org.zalando.nakadi.domain.DefaultStorage;
 import org.zalando.nakadi.domain.Storage;
 import org.zalando.nakadi.exceptions.runtime.NoStorageException;
 import org.zalando.nakadi.exceptions.runtime.StorageIsUsedException;
 import org.zalando.nakadi.repository.db.StorageDbRepository;
+import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
 import org.zalando.nakadi.utils.TestUtils;
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -25,7 +27,8 @@ public class StorageServiceTest {
     @Before
     public void setUp() {
         storageDbRepository = mock(StorageDbRepository.class);
-        storageService = new StorageService(TestUtils.OBJECT_MAPPER, storageDbRepository);
+        storageService = new StorageService(TestUtils.OBJECT_MAPPER, storageDbRepository,
+                new DefaultStorage(mock(Storage.class)), mock(ZooKeeperHolder.class));
     }
 
     @Test